Added basic caching capabilities
Most times information about "whatever" can be safely drawn from a
cache, especially for players in isolationist mode. This patch adds
a basic caching functionality through the Games::Lacuna::Client::Cache
module.

GLC::Cache manages a simple cache in a single file, saved with
Storable::nstore and loaded with Storable::retrieve. The cache is
made up of a three-level HoH structure where the first level is
indexed by url, the second level is indexed by method and the third
level is indexed by joining together all the parameters (separated with
"\x0"). The cache can be accessed with the retrieve() method (accepting the
same parameters as the call() method in GLC::RPC) and stuff can be put
with the store() method.

GL::Client has a new member variable "cache" (along with the accessor)
that points to the cache object. At the moment it is set as a new
instance of GLC::Cache. It could be anything supporting the retrieve() and
store() methods, anyway.

GLC::Module has been extended to use the cache in various steps if
it is available and it is requested. Cache can be ignored both in reading
and in writing, e.g. if it is known that a datum is no more valid
caching can be disabled for reading the info but could be enabled to
store the new stuff got from GLC::RPC.

If this sounds good, documentation will follow :)